switch "hlua" from using helgrind to "drd", the data race detector.
Adopt some best practices from the DRD manual:
http://valgrind.org/docs/manual/drd-manual.html

Wire up the lua_lock/unlock primitives with an actual lock implementation,
to make mark/trace of a thread stack race free.

The only issue that I can find so far is that the lock/unlock pairs need
protection for when errors are thrown; we can end up with mismatched locks.

In particular, the 301-basic.lua test reports destruction of a locked writer
lock on a particular table.

refs #5884 (spent 1.5)