Turned out that the majority DRD complaints were due to code generated with --enable-coverage is turned on.
Fixup the eq operator in the VM for strings; it would only succeed in
comparing two identical strings if they were interned at the same time or on
the same thread.

Add thread.mutex() as a way to create mutex objects; these can be created as
normal, recursive or error checking mutexes (defaults to error checking).

Add lua_pushcclosure2() as a way to create a C closure with an associated
static string.  At the moment this is used only by the gdb macros in the newly
added .gdbinit file, but can be added to the debug API and lua-visible stack
trace routines.

Added file/line information to the try/catch support so that it is easier to
figure out where things are in gdb.

refs #5884 (spent 3)