Default to a sane LUA_ROOT when built locally and not via configure. Don't trigger a GC run via lua_gc() or the lua-user-space collectgarbage routine.
Invoking the collector from any threads can lead to a deadlock and
should not be needed, as the collector runs every few seconds anyway.

Remove the CLOCK_MONOTONIC bits; without timer resolution compensation, this
effectively makes us sleep for "10 ticks" instead of 10 seconds, and the
timing just isn't that critical.

refs #5884 (spent 0.1) # time already accounted for on ticket