implement a simple, portable, stop-the-world mechanism on top of pthreads.
Ultimately, I'd like to replace this with lower-level suspend/resume
capabilities on the underlying OS: Solaris and Darwin have APIs that can be
used for this.  I think Linux has something at the syscall level that we may be
able to use.

I'd like to avoid SIGSTOP because it will be caught and noticed by parents of
the process that run lua; I'm mostly concerned about future bad interaction
with Gimli, which treats SIGSTOP as a signal to trace the child.

refs #5884 (spent 1.5)