Fix lpeg test: problem was with weak table handling.
Eliminate the use of malloc from the global collection routine; now we
allocate (or reallocate) an array of pthread_t each time a new thread is
discovered.

Add an atexit handler with the aim to getting valgrind to claim that we are
leak free.  Currently difficult due to libedit.

Also, we appear to mysteriously forget about io.stdin when running the test
harness, leaking 72 bytes.

refs #5884 (2.5)