adjust thread test to show (via output, not yet via assertions) that the threads are collected before the end of the script.
Adjusted the OS-thread and coroutine completion code to allow the threads to be collected sooner.

Worth noting the that lua_newthread API (used by the coroutine bits) will
always create an xref to the new thread, requiring a global trace before it can
be cleaned up.  The lua_newthreadref API doesn't associate the new thread with
the creating context (by pushing it on the stack), so it does not create an
xref.

This latter mode will be preferred in our integration with Momentum, as we can
then create independently collected lua_States per session and not require a
global collection cycle to guarantee reclamation of any critically important
boxed objects (such as messages).

refs #5884 (spent 1.25)