Fixed a logic error with the string interning; during collection we would unconditionally flush the contents of the string table, but we would not reset the size.
This meant that we would size the string table based on the total number of
strings that we had ever seen, which is visibly more memory than is strictly
needed (>20MB in reject.t).

In fixing this, it was noted that since we run a collection cycle after each
session, we were not reaping the benefits of string interning, so I changed
the logic to flush out the string table only if it contains more than 1024
strings.  This logic is fairly simplistic and we probably should look at
having some logic to remove strings larger than a certain size in this step,
but it changed a simple rejection policy script fairly dramatically:

Before:

{{{
52093 finished in state reading rcptto (868/s 52080/m 3124800/hr)
}}}

Fixed string interning:

{{{
112332 finished in state reading rcptto (1872/s 112320/m 6739200/hr)
}}}

refs #5884 (spent 1.5)