Add in thread.condition facility for creating condition variables. This is modeled after the python condition variable facility.
In so doing, discovered that the signal based stop-the-world system used
by the collected could encounter deadlocks with condition variables on
some systems, because POSIX forbids an EINTR return from
pthread_cond_wait, and requires that pthread_cond_wait re-acquire the
mutex before it returns.  This can lead to a deadlock if a different
thread currently owns the mutex and we are in the middle of a global
trace.

So, do away with the signal based scheme and adopt something similar to
the technique we use in block_collector(), which is to publish an intent
to stop and ensure that all write barriers respect this intent.

As a bonus, this approach allows all blocked mutators to share the
global trace workload and run concurrently.

Add a unit test for the condition functionality; it implements a simple
work queue entirely in lua script.