Introduce an API that simplifies building C-library bindings.
In addition to making it easier to bind to C code, it allows disparate
modules and libraries to pass data around, provided they know the
metatable name used by the canonical lua module.

In plainer English, this means that a general binding module can be
provided as a utility in the core lua package, and that an application
can create or consume instances of that library type without having any
special knowledge of the lua bindings, other than the metatable name
they should use.