FAD-1181 #time 2h Do a few things here:
1. Remove the old "block collector using signals" code, as it hasn't been used
in years, and creates naming confusion.
2. Redo the 'non-signal-collector' code to not try to participate in the
global trace thread work.  The participation in that work is not a bad thing,
but the busy waiting when there is nothing to do is not good.
3. Instead, use a heavyweight rwlock to synchronize threads that happen to try
to enter a barrier while a global trace is going on.  This allows those
threads to actually sleep while the trace is going on, and then when they are
able to acquire the lock they can then enter the barrier and be on their way.
4. Make all of those settings runtime-configurable as opposed to compile-time
options.

We tried to enable logic like this (that instead stopped threads trying to
enter a barrier rather than stopping all threads regardless of what they are
doing) a few years ago, but had some bad results with Lua table locking.  We
haven't changed any specific code there, but we have fixed a lot of problems
with things that needed to be in write barriers that in fact were not (we did
this as part of the Momentum 4 hardening).